feat: move CliHeader to primitives traits and rename to HeaderMut
